# docs/agent_os_openapi.yaml
openapi: 3.1.0
info:
  title: Enstack Agent OS API
  version: 1.0.0-draft.1
  summary: Give your AI agents real money — and real control.
  description: |
    # Enstack Agent OS — /v1 API

    The Agent OS API lets a verified **Owner** (a person or business) provision
    **Agents** — AI sub-principals that hold a Balance, spend on a Visa card,
    receive into a virtual bank account, and move money over local and
    cross-border rails via **Enstack Routing** — all inside an Owner-defined
    **Policy** (limits, allowlists, human approvals, full audit).

    ## Identity model
    - **Owner** — the KYC/KYB-verified principal. Owns all funds and all liability.
    - **Agent** — a sub-principal under an Owner. No separate KYC. Every agent
      action is attributable to both the agent and its owner.
    - **Credential** — a scoped API key or MCP token bound to exactly one agent
      (or to the owner). Credentials carry scopes; scopes are further constrained
      by the agent's Policy at execution time.
    - **Policy** — the authorization layer: per-transaction / daily / monthly
      limits, counterparty and category allowlists, and human-in-the-loop
      approval rules. Policy is versioned; every evaluation is audited.

    ## Conventions
    - **Versioning**: URL prefix `/v1/...`. Breaking changes ship as `/v2`.
    - **Idempotency**: `Idempotency-Key` header is **required on every mutating
      request** (POST/PUT/PATCH). Window: 24h, keyed by `(credential, key)`.
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /quotes:
    post:
      operationId: createQuote
      tags: [Quotes]
      summary: Create a quote
      description: |
        Prices a payment: source amount/currency → destination amount/currency
        over the best route Enstack Routing can offer. Quotes are firm until
        `expires_at`. Pass `quote_id` into payment creation to execute at the
        quoted rate.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteCreate' }
      responses:
        '201':
          description: Quote created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Quote' }
        default: { $ref: '#/components/responses/Error' }
  /quotes/{quote_id}:
    get:
      operationId: getQuote
      tags: [Quotes]
      summary: Retrieve a quote
      parameters:
        - name: quote_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The quote.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Quote' }
        default: { $ref: '#/components/responses/Error' }

  # ──────────────────────────── Payments ────────────────────────────
...
  /agents/{agent_id}/virtual_accounts:
    parameters:
      - $ref: '#/components/parameters/agentId'
    post:
      operationId: createVirtualAccount
      tags: [Virtual Accounts]
      summary: Create a virtual account for an agent
      description: |
        Provisions a named virtual bank account whose inbound credits post to
        the agent's balance. Account naming follows the scheme
        `{owner_legal_name} — {agent_display_name}` subject to banking-partner
        naming rules.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VirtualAccountCreate' }
      responses:
        '201':
          description: Virtual account created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VirtualAccount' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listVirtualAccounts
      tags: [Virtual Accounts]
      summary: List an agent's virtual accounts
      responses:
        '200':
          description: Virtual account list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/VirtualAccount' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /virtual_accounts/{virtual_account_id}:
    get:
      operationId: getVirtualAccount
      tags: [Virtual Accounts]
      summary: Retrieve a virtual account
      parameters:
        - name: virtual_account_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The virtual account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VirtualAccount' }
        default: { $ref: '#/components/responses/Error' }

  # ──────────────────────────── Approvals ────────────────────────────
...
  /webhook_endpoints:
    post:
      operationId: createWebhookEndpoint
      tags: [Webhooks]
      summary: Register a webhook endpoint
      description: |
        Owner scope only. Deliveries are HMAC-SHA256 signed
        (`Enstack-Signature` header), retried with exponential backoff, and
        replayable from the console. Payload schemas are defined in the
        companion **AsyncAPI** specification.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookEndpointCreate' }
      responses:
        '201':
          description: Endpoint registered. `secret` is returned only in this response.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookEndpointWithSecret' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listWebhookEndpoints
      tags: [Webhooks]
      summary: List webhook endpoints
      responses:
        '200':
          description: Endpoint list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookEndpoint' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /webhook_endpoints/{webhook_endpoint_id}:
    delete:
      operationId: deleteWebhookEndpoint
      tags: [Webhooks]
      summary: Delete a webhook endpoint
      parameters:
        - name: webhook_endpoint_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '204':
          description: Deleted.
        default: { $ref: '#/components/responses/Error' }
  /webhook_endpoints/{webhook_endpoint_id}/rotate_secret:
    post:
      operationId: rotateWebhookSecret
      tags: [Webhooks]
      summary: Rotate a webhook signing secret
      description: Old secret remains valid for 24h to allow zero-downtime rotation.
      parameters:
        - name: webhook_endpoint_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: New secret (returned only in this response).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookEndpointWithSecret' }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Owner key (`ok_live_…`) or agent credential (`ac_live_…`).
        Agent credentials are policy-bound: server-side Policy evaluation
        applies to every call regardless of credential scopes.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (UUIDv4 recommended). Required on every
        mutating request. 24h dedup window keyed by (credential, key).
        Same key + same body replays the original response; same key +
        different body returns 409.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    agentId:
      name: agent_id
      in: path
      required: true
      schema: { type: string }
      description: Agent identifier, e.g. `agt_01J….`
...
  responses:
    Error:
      description: |
        Error envelope. Notable codes: `policy_denied` (403, hard policy
        block), `insufficient_funds` (409), `idempotency_conflict` (409,
        key reuse with different body), `quote_expired` (409),
        `approval_required` is NOT an error — held actions return 202.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples: [policy_denied, insufficient_funds, quote_expired, idempotency_conflict, not_found, unauthorized, validation_error, rate_limited]
        message:
          type: string
          description: Human-readable description.
        details:
          type: [object, 'null']
          description: Field-level or context detail (e.g. which policy rule denied).
          additionalProperties: true
        request_id:
          type: string
          description: Correlate with support and the activity feed.
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d+)?$'
          description: Decimal string. Never a float.
          examples: ["1500.00", "0.4275"]
        currency:
          type: string
          description: ISO 4217 code.
          examples: [USD, PHP, HKD, EUR, SGD]
...
    QuoteCreate:
      type: object
      required: [agent_id, destination_currency]
      properties:
        agent_id: { type: string }
        source_amount:
          $ref: '#/components/schemas/Money'
          description: Provide source_amount OR destination_amount, not both.
        destination_amount: { $ref: '#/components/schemas/Money' }
        destination_currency: { type: string }
        destination_country:
          type: string
          description: ISO 3166-1 alpha-2. Improves route selection.
        counterparty_id:
          type: string
          description: If known, pins routing to rails that can reach this counterparty.
    Quote:
      type: object
      required: [id, agent_id, source_amount, destination_amount, rate, fees, expires_at, created_at]
      properties:
        id: { type: string, examples: [qot_01JABCDXYZ] }
        agent_id: { type: string }
        source_amount: { $ref: '#/components/schemas/Money' }
        destination_amount: { $ref: '#/components/schemas/Money' }
        rate:
          type: string
          description: Decimal string, destination units per source unit, all-in.
        fees:
          type: array
          items:
            type: object
            required: [type, amount]
            properties:
              type: { type: string, examples: [routing_fee, fx_spread, rail_fee] }
              amount: { $ref: '#/components/schemas/Money' }
        estimated_arrival:
          type: string
          description: Human estimate, e.g. "instant", "within 2 hours", "1–2 business days".
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...
    VirtualAccountCreate:
      type: object
      properties:
        currency:
          type: string
          description: Requested account currency. Availability varies by market.
        label: { type: string, maxLength: 60 }
    VirtualAccount:
      type: object
      required: [id, agent_id, account_name, status, created_at]
      properties:
        id: { type: string, examples: [va_01JABCDXYZ] }
        agent_id: { type: string }
        label: { type: [string, 'null'] }
        account_name: { type: string }
        account_number: { type: [string, 'null'] }
        bank_name: { type: [string, 'null'] }
        bank_identifier: { type: [string, 'null'] }
        currency: { type: string }
        supported_rails:
          type: array
          items: { type: string }
        status: { type: string, enum: [provisioning, active, closed] }
        created_at: { type: string, format: date-time }
...
    WebhookEndpointCreate:
      type: object
      required: [url, event_types]
      properties:
        url: { type: string, format: uri }
        event_types:
          type: array
          minItems: 1
          items: { type: string }
          description: Dot-notation types or prefixes, e.g. `payment.*`, `approval.requested`.
        label: { type: string, maxLength: 120 }
    WebhookEndpoint:
      type: object
      required: [id, url, event_types, status, created_at]
      properties:
        id: { type: string, examples: [whe_01JABCDXYZ] }
        url: { type: string, format: uri }
        event_types:
          type: array
          items: { type: string }
        label: { type: [string, 'null'] }
        status: { type: string, enum: [active, disabled] }
        created_at: { type: string, format: date-time }
    WebhookEndpointWithSecret:
      allOf:
        - $ref: '#/components/schemas/WebhookEndpoint'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: HMAC-SHA256 signing secret. Shown once.